% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_codes.R
\name{get_species_codes}
\alias{get_species_codes}
\title{Available species, codes, and assignments in \pkg{woodValuationDE}}
\usage{
get_species_codes()
}
\value{
A list with the species, species codes, and assignments to economic
        species groups available in \pkg{woodValuationDE}.
}
\description{
The function shows the availabe species, species codes and species
assignments to groups for the economic valuation in order to inform the users
for their own applications.
}
\examples{
get_species_codes()
}
