% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_codes.R
\name{get_species_codes}
\alias{get_species_codes}
\title{Available species in \pkg{woodValuationDE}, their codes, and parameter
assignments}
\usage{
get_species_codes(method = "fuchs.orig")
}
\arguments{
\item{method}{argument that is currently not used, but offers the possibility
to implement alternative parameters and functions in the
future.}
}
\value{
A list with the species, species codes, and assignments to economic
        species groups available in \pkg{woodValuationDE}.
}
\description{
The function returns the available species, species codes, and assignments of
species to species groups for the economic valuation.
}
\examples{
get_species_codes()
}
