% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wood_revenues.R
\name{wood_revenues}
\alias{wood_revenues}
\title{Wood revenues per unit volume}
\usage{
wood_revenues(
  diameter.q,
  species,
  value.level = 2,
  process.type = "combined",
  price.ref.assortment = "baseline",
  calamity.type = "none",
  calamity.factors = "baseline",
  species.code.type = "en"
)
}
\arguments{
\item{diameter.q}{Quadratic mean of the diameter at breast height (dbh) of
the harvested trees \eqn{[cm]}{[cm]}.}

\item{species}{Tree species, using an available \code{species.code.type}. For
a list with the available species and codes call
\code{\link{get_species_codes}}.}

\item{value.level}{Stand quality expressed as integer of \code{1:3}, with
\code{1} for an extraordinary high stand quality with high
shares of wood suitable for furniture, \code{2} for an
average quality, and \code{3} for an extraordinary low
quality (e.g. trees with many thick branches or stands
with massive ungulate damages). The value.levels refer to
the applied assortment tables (Offer and Staupendahl,
2018).}

\item{process.type}{Type of harvest process, with \code{"manually"} for
motor-manual harvest using a chain saw,
\code{"harvester"} for highly mechanized harvest
machines, or \code{"combined"} for a combination of the
previous types dependent on the mean diameter.}

\item{price.ref.assortment}{Wood price of the reference assortments allowing
to consider market fluctuations. Default is
\code{"baseline"} referring to the prices from 2010 to
2015 in Hesse, Germany (for details see readme of
\pkg{woodValuationDE} or von Bodelschwingh (2018)).
Alternatively, it can be user-provided tibble with the
same structure. Column species uses the specified
\code{species.code.type}.}

\item{calamity.type}{Type of a potential calamity determining the applied
calamity corrections, which implement reduced returns
and higher harvest costs. By default no calamity is
assumed \code{"none"}, \code{"calamity.dieter.2001"}
refers to a general larger calamity applying the
corrections according to Dieter (2001), \code{"ips"}
refers to quality losses due to infestations by the
European spruce bark beetle or \code{"ips.timely"} for
timely salvage fellings in less advanced attack stages
(both according to Fuchs et al. under review), and
\code{"stand.damage.fuchs"} to damages affecting only
one stand, \code{"regional.calamity.fuchs"} to
calamities with regional market effects and
\code{"national.calamity.fuchs"} to calamities affecting
(inter)national wood markets (the last three referring
to Fuchs et al. in preparation). User-defined types can
be implemented via the \code{calamity.factors}
parameter.}

\item{calamity.factors}{Summands \eqn{[EUR m^{-3}]}{[EUR m^(-3)]}
and factors to consider the consequences of
calamities on wood revenues and harvest costs.
\code{"baseline"} provides a tibble based on the
references listed in \code{calamity.type} (for
details see readme of \pkg{woodValuationDE}).
Alternatively, it can be user-provided tibble with
the same structure.}

\item{species.code.type}{Type of code in which \code{species} is given.
\code{"en"} for English species names or
\code{"nds"} for numeric species codes used in Lower
Saxony, Germany. For a list with the available
species and codes call
\code{\link{get_species_codes}}.}
}
\value{
A vector with wood revenues per unit volume
        \eqn{[EUR m^{-3}]}{[EUR m^(-3)]}. The volume refers to the salable
        wood volume, provided by \code{\link{vol_salable}}.
}
\description{
The function estimates wood revenues per unit volume based on the assortment
tables of Offer and Staupendahl (2018) and the wood revenue function of von
Bodelschwingh (2018). Consequences of calamities are implemented based on
Dieter (2001), Fuchs et al. (under review) and Fuchs et al. (in preparation).
Apart from Dieter (2001), all functions and factors are based on data of
HessenForst, the forest administration of the Federal State of Hesse in
Germany. For further details see the \pkg{woodValuationDE} readme.
}
\examples{
wood_revenues(40,
              "beech")

# species codes Lower Saxony (Germany)
wood_revenues(40,
              211,
              species.code.type = "nds")

# vector input
wood_revenues(seq(20, 50, 5),
              "spruce")

wood_revenues(40,
              rep(c("beech", "spruce"),
                  each = 3),
              value.level = rep(1:3, 2))

wood_revenues(40,
              rep("spruce", 7),
              calamity.type = c("none",
                                "calamity.dieter.2001",
                                "ips",
                                "ips.timely",
                                "stand.damage.fuchs",
                                "regional.calamity.fuchs",
                                "national.calamity.fuchs"))

# user-defined calamities with respective changes in wood revenues
wood_revenues(40,
              rep("spruce", 3),
              calamity.type = c("none",
                                "my.own.calamity.1",
                                "my.own.calamity.2"),
              calamity.factors = dplyr::tibble(
                calamity.type = rep(c("none",
                                      "my.own.calamity.1",
                                      "my.own.calamity.2"),
                                    each = 2),
                species.group = rep(c("softwood",
                                      "deciduous"),
                                    times = 3),
                revenues.factor = c(1.0, 1.0,
                                    0.8, 0.8,
                                    0.2, 0.2),
                cost.factor = c(1.0, 1.0,
                                1.5, 1.5,
                                1.0, 1.0),
                cost.additional = c(0, 0,
                                    0, 0,
                                    5, 5)))

# adapted market situation by providing alternative prices for the reference assortments
wood_revenues(40,
              c("oak", "beech", "spruce"))
wood_revenues(40,
              c("oak", "beech", "spruce"),
              price.ref.assortment = dplyr::tibble(
                species = c("oak", "beech", "spruce"),
                price.ref.assortment = c(300, 80, 50)))

}
\references{
Dieter, Matthias (2001): Land expectation values for spruce and
            beech calculated with Monte Carlo modelling techniques. For.
            Policy Econ. 2 (2), S. 157-166.
            \doi{10.1016/S1389-9341(01)00045-4}.

Fuchs, Jasper M., Hittenbeck, Anika, Brandl, Susanne, Schmidt,
            Matthias, Paul, Carola (under review): Adaptation Strategies for
            Spruce Forests - Economic Potential of Bark Beetle Management and
            Douglas Fir Cultivation in Future Tree Species Portfolios.

Fuchs, Jasper M., von Bodelschwingh, Hilmar, Paul, Carola,
            Husmann, Kai (in preparation): Applying Time Series Analysis to
            Quantify the Impact of Quality and Supply Changes After
            Disturbances on Wood Revenues.

Offer, Armin and Staupendahl, Kai (2018): Holzwerbungskosten- und
            Bestandessortentafeln (Wood Harvest Cost and Assortment
            Tables). Kassel: HessenForst (publisher).

von Bodelschwingh, Hilmar (2018): Oekonomische Potentiale von
            Waldbestaenden. Konzeption und Abschaetzung im Rahmen einer
            Fallstudie in hessischen Staatswaldflaechen (Economic Potentials
            of Forest Stands and Their Consideration in Strategic Decisions).
            Bad Orb: J.D. Sauerlaender`s Verlag (Schriften zur Forst- und
            Umweltoekonomie, 47).
}
